components/analytics-service-dev/pyproject.toml
[project]
name = "analytics-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiobotocore>=3.1.0",
    "boto3>=1.42.19",
    "clickhouse-connect>=0.10.0",
    "dill>=0.4.0",
    "fastapi>=0.128.0",
    "faststream[rabbit]>=0.6.5",
    "loguru>=0.7.3",
    "msgspec>=0.20.0",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "tomlkit>=0.14.0",
    "ujson>=5.11.0",
    "uvicorn>=0.40.0",
]


[tool.ruff]
line-length = 100
extend-exclude = [
    "src/migrations",
]
force-exclude = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # isort
    "I",
    # tidy imports
    "TID",
    # flake8-bugbear
    "B",
    # flake8-bandit
    "S",
    # pydocstyle
    "D",
    # pyugrade
    "UP"
]
ignore = ["D104", "D107", "B008", "E721"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101", "D100", "D101", "D102", "D103", "F403", "D106"]

[tool.ruff.lint.pycodestyle]
max-doc-length = 100
max-line-length = 100

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
docstring-code-format = true


[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.0",
    "ruff>=0.11.13,<=0.12.0",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.8",
    "testcontainers>=4.13.3",
    "pytest-dotenv>=0.5.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
